mysite/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
psycopg2==2.9.1
pydub==0.25.1